+++ src/main/java/site/siredvin/gttruesteam/common/Constants.java
@@ -4,4 +4,10 @@ public class Constants {
 
     public static int FLUID_INGOT = 144;
     public static int FLUID_BLOCK = FLUID_INGOT * 9;
+    public static int BASE_STEAM_TEMPERATURE = 373;
+    public static double BASE_STEAM_DENSITY = 0.5;
+    public static int TEMPERATURE_TO_ENERGY = 50;
+    public static int STEAM_COMPRESSION_COEF = 20;
+    public static int STEAM_CRITICAL_BOOST = 2;
+    public static int METAPHYSICS_STEAM_BOILING_OUTPUT = 6000;
 }